Draft of refactoring avocado-tests.
